refactor: Move feature groups into better location
