+++ mloda_plugins/feature_group/input_data/read_dbs/sqlite.py
@@ -8,7 +8,7 @@ from mloda_core.abstract_plugins.components.feature_set import FeatureSet
 from mloda_core.abstract_plugins.components.hashable_dict import HashableDict
 from mloda_core.abstract_plugins.components.data_types import DataType
 from mloda_core.abstract_plugins.components.options import Options
-from mloda_plugins.input_data.read_db import ReadDB
+from mloda_plugins.feature_group.input_data.read_db import ReadDB
 
 
 class SQLITEReader(ReadDB):
