+++ tests/test_plugins/feature_group/test_input_data/test_read.py
@@ -4,6 +4,7 @@ from typing import Any, List, Optional
 import tempfile
 import unittest
 import sqlite3
+from mloda_plugins.feature_group.input_data.read_files.csv import CsvReader
 import pytest
 import pyarrow as pa
 
@@ -15,10 +16,9 @@ from mloda_core.abstract_plugins.components.index.index import Index
 from mloda_core.abstract_plugins.components.link import Link
 from mloda_core.abstract_plugins.components.plugin_option.plugin_collector import PlugInCollector
 from mloda_core.api.request import mlodaAPI
-from mloda_plugins.input_data.read_dbs.sqlite import SQLITEReader
-from mloda_plugins.input_data.read_file_feature import ReadFileFeature
-from mloda_plugins.input_data.read_files.csv import CsvReader
-from tests.test_core.test_input_data.test_classes.test_input_classes import (
+from mloda_plugins.feature_group.input_data.read_dbs.sqlite import SQLITEReader
+from mloda_plugins.feature_group.input_data.read_file_feature import ReadFileFeature
+from tests.test_plugins.feature_group.test_input_data.test_classes.test_input_classes import (
     DBInputDataTestFeatureGroup,
     ReadFileFeatureWithIndex,
 )
