+++ tests/test_plugins/feature_group/test_input_data/test_read_dbs/test_sqlite.py
@@ -8,7 +8,7 @@ import pyarrow as pa
 from mloda_core.abstract_plugins.components.feature import Feature
 from mloda_core.abstract_plugins.components.feature_set import FeatureSet
 from mloda_core.abstract_plugins.components.hashable_dict import HashableDict
-from mloda_plugins.input_data.read_dbs.sqlite import SQLITEReader
+from mloda_plugins.feature_group.input_data.read_dbs.sqlite import SQLITEReader
 
 
 class MockFeatureSet:
